// projects/ctng/wordpress/src/lib/model/wp-image.ts
export interface WPImage {
  ID: number;
  id: number;
  alt: string;
  author: number;
  description: string;
  filename: string;
  filesize: number;
  height: number;
  width: number;
  icon: string;
  link: string;
  date: string;
  modified: string;
  name: string;
  subtype: string;
  title: string;
  type: string;
  uploaded_to: number;
  url: string;
  caption: string;
  sizes: any;
}